requirements_dev.txt
amqp==5.0.5
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
httplib2==0.19.0
idna==2.10
kombu==5.0.2
oauth2==1.9.0.post1
Pillow==8.1.2
prompt-toolkit==3.0.17
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
